Replaced Parse::RecDescent parser with Mitchell's parser. Changed Inline::C so that both parsers can be called the same way.
Wrote a test suite that compares the output of both parsers on a
set of source code snippets.

We need to add a slew of test cases to the suite. If they all pass
we can release this parser as the default.

We need to add a PARSER config option to allow the user to change
parsers.


git-svn-id: http://svn.kwiki.org/ingy/cpan/trunk/Inline@106 9ffafc32-93ef-0310-a7a6-903af07b6741