Fix failing tests on systems missing perldiag.pod
Replaced 'use diagnostics' with an eval. This way tests will continue
on systems where perldiag.pod (that diagnostics depend on) is missing.
This is most notably on Cygwin, where POD files are in a separate
package called perl_pods.

This is part of the CPAN Pull Request Challenge, where I got Inline
as my March assignment.