Relax the version requirement for installing modules
Inline and Inline::MakeMaker both want the VERSION of the module
to match #.##, but this isn't convenient for developer releases with '_'.
(and previously, they disagreed on whether additional digits were allowed)

There doesn't seem to be any technical limitation on the underscore
numbers, so relax the regexes and update the error messages.

Fixes #62

Also allow multi-digit major version - #62