Fix race condition between file creation and locking
Because creating the configuration file does not atomically lock it, it
is possible for a reader to see that a configuration file exists, open
it, lock it, and read it before the writer can acquire an exclusive lock
and fill the file. The reader will then die with a misleading error
message ("You are using a config file that was created by an older
version of Inline") because at that point the file is still empty.

By treating an empty configuration file as absent, we ensure that
*someone* (either ourselves or a concurrent writer) will fill the
configuration file before we proceed to read it.